working happy path Adaptor test
