+++ src/SFA.DAS.Funding.SystemAcceptanceTests/SFA.DAS.Funding.SystemAcceptanceTests.csproj
@@ -18,8 +18,8 @@
     <PackageReference Include="RestAssured.Net" Version="1.0.0" />
     <PackageReference Include="SFA.DAS.Apprenticeships.Types" Version="1.1.4" />
 	<PackageReference Include="SFA.DAS.CommitmentsV2.Messages" Version="100.0.17" />
-    <PackageReference Include="SFA.DAS.Funding.ApprenticeshipEarnings.Types" Version="1.0.18-prerelease-6" />
-	<PackageReference Include="SFA.DAS.Funding.ApprenticeshipPayments.Types" Version="0.1.8-prerelease-78" />
+    <PackageReference Include="SFA.DAS.Funding.ApprenticeshipEarnings.Types" Version="1.0.18-prerelease-7" />
+	<PackageReference Include="SFA.DAS.Funding.ApprenticeshipPayments.Types" Version="0.1.8-prerelease-92" />
     <PackageReference Include="SpecFlow.Plus.LivingDocPlugin" Version="3.9.57" />
     <PackageReference Include="SpecFlow.NUnit" Version="3.9.74" />
     <PackageReference Include="nunit" Version="3.13.3" />
