+++ src/SFA.DAS.Funding.SystemAcceptanceTests/StepDefinitions/CalculateUnfundedPaymentsStepDefinitions.cs
@@ -16,6 +16,7 @@ namespace SFA.DAS.Funding.SystemAcceptanceTests.StepDefinitions
         private List<FinalisedOnProgammeLearningPaymentEvent> _finalisedPaymentsList;
         private TestSupport.Payments[] _paymentEntity;
         private byte _currentCollectionPeriod;
+        private const int SecondToWaitBeforePublishingReleasePaymentsCommandToSimulateEndOfMonth = 30;
 
         public CalculateUnfundedPaymentsStepDefinitions(ScenarioContext context)
         {
@@ -59,12 +60,14 @@ namespace SFA.DAS.Funding.SystemAcceptanceTests.StepDefinitions
         [When(@"the scheduler triggers Unfunded Payment processing")]
         public async Task SchedulerTriggersUnfundedPaymentProcessing()
         {
+            Thread.Sleep(TimeSpan.FromSeconds(SecondToWaitBeforePublishingReleasePaymentsCommandToSimulateEndOfMonth));
             await _paymentsMessageHelper.PublishReleasePaymentsCommand(_releasePaymentsCommand);
         }
 
         [When(@"the Release Payments command is published again")]
         public async Task WhenTheReleasePaymentsCommandIsPublishedAgain()
         {
+            Thread.Sleep(TimeSpan.FromSeconds(SecondToWaitBeforePublishingReleasePaymentsCommandToSimulateEndOfMonth));
             FinalisedOnProgrammeLearningPaymentEventHandler.ReceivedEvents.Clear();
 
             await _paymentsMessageHelper.PublishReleasePaymentsCommand(_releasePaymentsCommand);
